+++ pkg/client/cli/cmd_status.go
@@ -53,14 +53,18 @@ func daemonStatus(cmd *cobra.Command) error {
 			return err
 		}
 
+		dns := status.OutboundConfig.Dns
 		fmt.Fprintln(out, "Root Daemon: Running")
 		fmt.Fprintf(out, "  Version   : %s (api %d)\n", version.Version, version.ApiVersion)
 		fmt.Fprintf(out, "  DNS       :\n")
-		fmt.Fprintf(out, "    Local IP        : %v\n", net.IP(status.OutboundConfig.Dns.LocalIp))
-		fmt.Fprintf(out, "    Remote IP       : %v\n", net.IP(status.OutboundConfig.Dns.RemoteIp))
-		fmt.Fprintf(out, "    Exclude suffixes: %v\n", status.OutboundConfig.Dns.ExcludeSuffixes)
-		fmt.Fprintf(out, "    Include suffixes: %v\n", status.OutboundConfig.Dns.IncludeSuffixes)
-		fmt.Fprintf(out, "    Timeout         : %v\n", status.OutboundConfig.Dns.LookupTimeout.AsDuration())
+		if dns.LocalIp != nil {
+			// Local IP is only set when the overriding resolver is used
+			fmt.Fprintf(out, "    Local IP        : %v\n", net.IP(dns.LocalIp))
+		}
+		fmt.Fprintf(out, "    Remote IP       : %v\n", net.IP(dns.RemoteIp))
+		fmt.Fprintf(out, "    Exclude suffixes: %v\n", dns.ExcludeSuffixes)
+		fmt.Fprintf(out, "    Include suffixes: %v\n", dns.IncludeSuffixes)
+		fmt.Fprintf(out, "    Timeout         : %v\n", dns.LookupTimeout.AsDuration())
 		fmt.Fprintf(out, "  Also Proxy: (%d subnets)\n", len(status.OutboundConfig.AlsoProxySubnets))
 		for _, subnet := range status.OutboundConfig.AlsoProxySubnets {
 			fmt.Fprintf(out, "    - %s\n", iputil.IPNetFromRPC(subnet))